+++ GridKit/Model/PhasorDynamics/SystemModelData.hpp
@@ -15,6 +15,7 @@
 #include <GridKit/Model/PhasorDynamics/Exciter/ESDC1A/Esdc1aData.hpp>
 #include <GridKit/Model/PhasorDynamics/Exciter/IEEET1/Ieeet1Data.hpp>
 #include <GridKit/Model/PhasorDynamics/Exciter/SEXS-PTI/SexsPtiData.hpp>
+#include <GridKit/Model/PhasorDynamics/Governor/GASTPTI/GastPtiData.hpp>
 #include <GridKit/Model/PhasorDynamics/Governor/HYGOV/HygovData.hpp>
 #include <GridKit/Model/PhasorDynamics/Governor/Tgov1/Tgov1Data.hpp>
 #include <GridKit/Model/PhasorDynamics/Load/LoadZ/LoadZData.hpp>
@@ -49,6 +50,7 @@ namespace GridKit
       using RepcaDataT              = Controller::RepcaData<RealT, IdxT>;
       using Tgov1DataT              = Governor::Tgov1Data<RealT, IdxT>;
       using Esdc1aDataT             = Exciter::Esdc1aData<RealT, IdxT>;
+      using GastPtiDataT            = Governor::GastPtiData<RealT, IdxT>;
       using HygovDataT              = Governor::HygovData<RealT, IdxT>;
       using Ieeet1DataT             = Exciter::Ieeet1Data<RealT, IdxT>;
       using SexsPtiDataT            = Exciter::SexsPtiData<RealT, IdxT>;
@@ -112,6 +114,7 @@ namespace GridKit
       std::vector<LoadZIPDataT>            loadzip;         ///< LoadZIP instances within the model
       std::vector<Tgov1DataT>              gov;             ///< Governors within the model
       std::vector<Esdc1aDataT>             esdc1a;          ///< ESDC1A exciters within the model
+      std::vector<GastPtiDataT>            gastpti;         ///< GASTPTI governors within the model
       std::vector<HygovDataT>              hygov;           ///< HYGOV governors within the model
       std::vector<Ieeet1DataT>             exciter;         ///< Exciters within the model
       std::vector<SexsPtiDataT>            sexspti;         ///< SEXS-PTI exciters within the model