+++ tests/UnitTests/PhasorDynamics/runSystemSingleComponentTests.cpp
@@ -17,6 +17,7 @@ int main()
   result += test.loadZIP();
   result += test.regca();
   result += test.repca();
+  result += test.gastpti();
   result += test.genrou();
   result += test.genClassical();
   result += test.tgov1();